Fix getting max memory from env var

+++ R/functions/extrapolate_to_env.R
@@ -79,7 +79,7 @@ extrapolate_to_env <- function(
                                  new = seq.int(n_x_row)[nonsingular_det_sites])
 
 		mem_per_pair <- 4 * (6 + 4 * n_preds + 2 * n_preds^2)
-		if (is.na(mem_max <- as.integer(Sys.getenv("TENSOR_MEM_MAX", "")))) {
+		if (is.na(mem_max <- as.numeric(Sys.getenv("TENSOR_MEM_MAX", "")))) {
 				n_row_batch <- floor(mem_max / mem_per_pair)
 		} else {
 				n_row_batch <- floor(mem_max / mem_per_pair)
